+++ apps/cli-playground/eslint.config.js
@@ -1,0 +1,14 @@
+import { sheriff, tseslint } from 'eslint-config-sheriff';
+
+const sheriffOptions = {
+  react: true,
+  lodash: true,
+  remeda: true,
+  next: true,
+  astro: true,
+  playwright: true,
+  jest: true,
+  vitest: true,
+};
+
+export default tseslint(sheriff(sheriffOptions));