+++ apps/docs-website/src/utils/filterDuplicatedRules.ts
@@ -1,6 +1,6 @@
-import { isEqual } from "lodash-es";
-import type { Entry } from "@sherifforg/types";
-import { removeElementFromArrayByValue } from "./removeElementFromArrayByValue";
+import { isEqual } from 'lodash-es';
+import type { Entry } from '@sherifforg/types';
+import { removeElementFromArrayByValue } from './removeElementFromArrayByValue';
 
 const isSameItem = (firstOccurrence: Entry, secondOccurrence: Entry) => {
   return (