+++ packages/eslint-config-sheriff/src/handpickedRules/importHandPickedRules.ts
@@ -1,3 +1,5 @@
+import type { TSESLint } from '@typescript-eslint/utils';
+
 export const importHandPickedRules = {
   // https://github.com/import-js/eslint-plugin-import/issues/2913
   // https://github.com/import-js/eslint-plugin-import/issues/2673
@@ -9,4 +11,4 @@ export const importHandPickedRules = {
   'import/no-duplicates': [2, { 'prefer-inline': true }],
   'import/no-useless-path-segments': [2, { noUselessIndex: true }],
   // 'import/consistent-type-specifier-style': [2, 'prefer-inline'], // this is not actually needed when "@typescript-eslint/no-import-type-side-effects is set. Explanation here: https://github.com/import-js/eslint-plugin-import/issues/2676#issuecomment-1407107260
-};
+} as const satisfies TSESLint.FlatConfig.Rules;