+++ packages/eslint-config-sheriff/src/remedaConfig.ts
@@ -1,7 +1,7 @@
 import remedaPlugin from 'eslint-plugin-remeda';
-import { supportedFileTypes } from '@sherifforg/constants';
 import tseslint from 'typescript-eslint';
-import { type TSESLint } from '@typescript-eslint/utils';
+import { supportedFileTypes } from '@sherifforg/constants';
+import type { TSESLint } from '@typescript-eslint/utils';
 
 export const remedaConfig: TSESLint.FlatConfig.ConfigArray = tseslint.config({
   files: [supportedFileTypes],