+++ packages/sheriff-create-config/src/index.ts
@@ -1,11 +1,14 @@
 #!/usr/bin/env node
+/* eslint-disable sonarjs/no-duplicate-string */
 
 import { consola } from 'consola';
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
+import { CLIOptionsCatalog } from '@sherifforg/constants';
 import packageJson from '../package.json';
 import { askForCustomPath } from './utils/askForCustomPath';
 import { askForEslintTsConfig } from './utils/askForEslintTsConfig';
+import { askForPrettierInstallation } from './utils/askForPrettierInstallation';
 import { askForPrettierSupport } from './utils/askForPrettierSupport';
 import { setDependencies } from './utils/setDependencies';
 import { setEslintConfig } from './utils/setEslintConfig';
@@ -17,8 +20,21 @@ const { argv } = yargs(hideBin(process.argv))
   .version(packageJson.version)
   .alias('v', 'version')
   .option('filter', {
-    type: 'string',
-    description: 'Filter for specific workspace',
+    type: CLIOptionsCatalog.filter.type,
+    description: CLIOptionsCatalog.filter.description,
+  })
+  .option('typescript', {
+    type: CLIOptionsCatalog.typescript.type,
+    description: CLIOptionsCatalog.typescript.description,
+  })
+  .option('prettier', {
+    type: CLIOptionsCatalog.prettier.type,
+    description: CLIOptionsCatalog.prettier.description,
+  })
+  .option('install-deps', {
+    type: CLIOptionsCatalog['install-deps'].type,
+    description: CLIOptionsCatalog['install-deps'].description,
+    default: CLIOptionsCatalog['install-deps'].default,
   })
   .help()
   .alias('h', 'help');
@@ -26,23 +42,44 @@ const { argv } = yargs(hideBin(process.argv))
 // eslint-disable-next-line
 async function main() {
   const commandArguments = await argv;
+
+  /**
+   * When true, the command is being run in a package within a monorepo.
+   */
   const isWorkspace = Boolean(commandArguments.filter);
+  const isTypescriptPreRequested = commandArguments.typescript;
+  const isPrettierPreRequested = commandArguments.prettier;
+  const isDependenciesInstallationPreRequested =
+    commandArguments['install-deps'];
 
   showWelcome();
+
   const customProjectRootPath = isWorkspace ? await askForCustomPath() : null;
-  const hasLocalPrettierSupport = isWorkspace
-    ? await askForPrettierSupport()
-    : false;
-  const isEslintTsConfig = await askForEslintTsConfig();
+
+  const isEslintTsConfig =
+    isTypescriptPreRequested ?? (await askForEslintTsConfig());
 
   await setEslintConfig(isEslintTsConfig, customProjectRootPath);
 
-  if (!isWorkspace || hasLocalPrettierSupport) {
+  let shouldInstallPrettier = false;
+
+  const hasLocalPrettierSupport =
+    isPrettierPreRequested ?? (await askForPrettierSupport());
+
+  if (hasLocalPrettierSupport) {
+    shouldInstallPrettier = await askForPrettierInstallation();
     await setPrettierConfig(customProjectRootPath);
     await setPrettierIgnore(customProjectRootPath);
   }
-  await setDependencies(customProjectRootPath);
-  consola.info("You're all set!");
+
+  await setDependencies({
+    customProjectRootPath,
+    shouldInstallPrettier,
+    shouldInstallJiti: isEslintTsConfig,
+    shouldInstallDependencies: isDependenciesInstallationPreRequested,
+  });
+
+  consola.info("You're all set! Happy coding 🎉");
 }
 
 // eslint-disable-next-line